+++ pkg/analyzer_plugin/test/support/abstract_single_unit.dart
@@ -49,7 +49,7 @@ class AbstractSingleUnitTest extends AbstractContextTest {
   AstNode findNodeAtOffset(int offset, [Predicate<AstNode> predicate]) {
     AstNode result = new NodeLocator(offset).searchWithin(testUnit);
     if (result != null && predicate != null) {
-      result = result.getAncestor(predicate);
+      result = result.thisOrAncestorMatching(predicate);
     }
     return result;
   }