+++ pkg/analyzer/CHANGELOG.md
@@ -3,6 +3,8 @@
   option now has the combined semantics of both options. This means that users
   that disable `implicit-casts` might now see errors that were not previously
   being reported.
+* Deprecated `AstNode.getAncestor` and introduced
+  `AstNode.thisOrAncestorMatching` as its replacement.
 
 ## 0.33.4
 * Add a hint when either Future or Stream are imported from dart:core in a package that is expected to work with an SDK before 2.1 where they were required to be imported from dart:async.